requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
dj-database-url
python-dotenv
google-generativeai
groq